pyproject.toml
[project]
name = "fuseformer-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "Danial Ahangarani",email = "daniahangarani@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "matplotlib (==3.9.4)",
    "torchvision (>=0.21.0,<0.22.0)",
    "lightning (>=2.5.0.post0,<3.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "mlflow (>=2.20.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "fuseformer_poetry", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
